第一次作业

+++ W1-1/contracts/Counter.sol
@@ -12,7 +12,7 @@ contract Counter {
         }
 
 
-    function set(uint x) public {
+    function set(uint8 x) public {
         counter = counter + x;
         }
 }
